Add functional-test inference-pipline
