+++ tests/functional_tests/test_cases/inference-pipline/Cambricon_MLU/conf/Cambricon_MLU.yaml
@@ -1,0 +1,29 @@
+defaults:
+  - _self_
+  - inference: Cambricon_MLU
+
+experiment:
+  exp_name: Cambricon_MLU
+  exp_dir: tests/functional_tests/test_cases/inference-pipline/Cambricon_MLU/results_test/Cambricon_MLU
+  task:
+    type: inference
+    backend: vllm
+    entrypoint: flagscale/inference/inference_aquila.py
+  runner:
+    hostfile: null
+  cmds:
+    before_start:
+      source /root/miniconda3/bin/activate flagscale-inference
+  envs:
+    VLLM_LATENCY_DEBUG: true
+    CNCL_CNPX_DOMAIN_ENABLE: 0
+    PYTORCH_MLU_ALLOC_CONF: "expandable_segments:True"
+    EXPERT_PARALLEL_EN: true
+    VLLM_GRAPH_DEBUG: false
+    VLLM_AVG_MOE_EN: 1
+
+action: run
+
+hydra:
+  run:
+    dir: ${experiment.exp_dir}/hydra
