+++ tests/functional_tests/test_cases/inference-pipline/nvidia_flaggems/results_gold/nvidia_flaggems
@@ -1,0 +1,8 @@
+**************************************************
+output.prompt='The first President of the United States'
+output.outputs[0].text=', George Washington, was born on February 22, 173'
+output.outputs[0].token_ids=[11, 9857, 6515, 11, 572, 9223, 389, 7400, 220, 17, 17, 11, 220, 16, 22, 18]
+**************************************************
+output.prompt='The capital of France'
+output.outputs[0].text=' is Paris. The capital of Germany is Berlin. The capital of Italy is Rome'
+output.outputs[0].token_ids=[374, 12095, 13, 576, 6722, 315, 9856, 374, 19846, 13, 576, 6722, 315, 15344, 374, 21718]
