+++ tests/scripts/_gpu_check.sh
