+++ tests/scripts/unit_tests/test_coverage.sh
